+++ config/queue.php
@@ -63,10 +63,9 @@ return [
             'driver' => 'redis',
             'connection' => 'default',
             'queue' => env('REDIS_QUEUE', 'default'),
-            'retry_after' => 90,
+            'retry_after' => 100,
             'block_for' => null,
         ],
-
     ],
 
     /*